feat: migrated supportedLanguagesUtils to ts
