+++ src/utils/getTranslationEnginePreferences.ts
@@ -1,5 +1,5 @@
 import { SetTranslateWithFallbackEngines } from "../services/translateService.d";
-import { Engines } from "../types/settings.d";
+import { Engines, TranslationEngines } from "../types/settings.d";
 
 /**
  * By default use Engines that doesn't require an API key
@@ -9,8 +9,8 @@ const DEFAULT_ENGINES = [Engines.GOOGLE, Engines.BING, Engines.LIBRE_TRANSLATE];
 /**
  * Returns an array of translation engines to use based on the provided settings and CLI arguments. If no one is provided then the default engines are returned.
  * @param {Object} options - The options object.
- * @param {Array} options.settingsTranslationEngines - The array of translation engines specified in the settings file.
- * @param {string} options.cliArgEngine - The translation engine specified as a CLI argument.
+ * @param {TranslationEngines} options.settingsTranslationEngines - The array of translation engines specified in the settings file.
+ * @param {Engines} options.cliArgEngine - The translation engine specified as a CLI argument.
  * @returns {Array} - The array of translation engines to use.
  */
 const getTranslationEnginesToUse = ({
