--- src/utils/normalizeFilePath.js
@@ -1,30 +1,0 @@
-/**
- * Normalizes a file path based on the operating system.
- *
- * @param {string} pathString - The string representing the file path to be normalized.
- * @returns {string} A normalized file path.
- *
- * @throws {Error} Throws an error if no string is provided.
- *
- * @example
- * // On Windows:
- * normalizeFilePath('file:///C:/Users/pepe/Documents/jobus/blazer/misc/i18n-populator/test-configs/test-config.json')
- * // Returns 'C:/Users/pepe/Documents/jobus/blazer/misc/i18n-populator/test-configs'
- *
- * @example
- * // On Unix:
- * normalizeFilePath('file://home/pepe/Documents/jobus/blazer/misc/i18n-populator/test-configs/test-config.json')
- * // Returns 'home/pepe/Documents/jobus/blazer/misc/i18n-populator/test-configs'
- */
-function normalizeFilePath(pathString) {
-  if (typeof pathString !== "string" || pathString.length === 0)
-    throw new Error("No string provided");
-
-  const path = pathString.split("/").slice(0, -1).join("/");
-
-  return process.platform === "win32"
-    ? path.replace(/^file:\/\/\/?/, "")
-    : path.replace("file://", "");
-}
-
-export { normalizeFilePath };
